nativetext: Refactor to use a single glyph map instead of multiple files, change order of 2D array
Rather than [x][y], we now index by [y][x]. This is so that we can move lines up and down easier, without having to iterate over each character -- instead, we can just iterate over each row.

Also, we're now using a single image to represent all the glyphs, and we're cropping them out in memory at runtime. This is to reduce the disk load when loading, and thus, reduce startup times (for the theme).